Partial Success: Graveyard Portion
Succeed at properly debugging and correcting problem to get partial
victory on Graveyard printout.  Line 41 (if letterx == guessx) is the
crutial line.  The order of Letter and Guess matters a lot for what should
be printed.  The operand is also very important: I struggled with != and
(by mistake) a single =.  It was only after I changed all of the variables
of this Graveyard loop to match "exactly" with the Word loop that I found
my error: wrong operand meant wrong printout.  I still have further to go
to get this Graveyard portion the way I want, but I'm proud that I
successfully debugged my own problem!  It's crazy how a single character
can confuse the hell out of you!