Updates to Graveyard loop AND lesson about returns
return lesson: I try to combine the separate def's into one big loop.
This led to confusion, the crux of which surrounded "return" function and
value.  Jim helped out a lot with that!  Learning more abuot the solution
allowed me to fix my original file and get the Graveyard display and tally
to work correctly (for now). ;-)